+++ docs/html/search/all_15.js
@@ -8,8 +8,9 @@ var searchData=
   ['references_5',['References',['../index.html#autotoc_md9',1,'']]],
   ['reset_6',['reset',['../group__mcu__reset.html#gadff1ae9368523be49ce8ef497d4ca902',1,'reset:&#160;reset.h'],['../group__mcu__reset.html',1,'Reset']]],
   ['reset_2eh_7',['reset.h',['../reset_8h.html',1,'']]],
-  ['run_20unit_20tests_20from_20terminal_20on_20linux_8',['How to run unit tests from terminal on Linux',['../md_utils_2csv2hex_2README.html#autotoc_md13',1,'']]],
-  ['rxd_5fp16_5ftxd_5fp17_9',['RxD_P16_TxD_P17',['../group__uart1__shared.html#ggaed599bf26e2baa98c2d108a072d02337a21147eba9db31edf831bc27da595c909',1,'uart1_shared.h']]],
-  ['rxd_5fp30_5ftxd_5fp31_10',['RxD_P30_TxD_P31',['../group__uart1__shared.html#ggaed599bf26e2baa98c2d108a072d02337abb894b0d9862cbfcd47c453dba0b9c47',1,'uart1_shared.h']]],
-  ['rxd_5fp36_5ftxd_5fp37_11',['RxD_P36_TxD_P37',['../group__uart1__shared.html#ggaed599bf26e2baa98c2d108a072d02337a5d9bd9050fee81f933458cb44d02eed3',1,'uart1_shared.h']]]
+  ['row_5flength_8',['ROW_LENGTH',['../mem__dump_8c.html#aa3a19a2da15cad9e01e11db672840e57',1,'mem_dump.c']]],
+  ['run_20unit_20tests_20from_20terminal_20on_20linux_9',['How to run unit tests from terminal on Linux',['../md_utils_2csv2hex_2README.html#autotoc_md13',1,'']]],
+  ['rxd_5fp16_5ftxd_5fp17_10',['RxD_P16_TxD_P17',['../group__uart1__shared.html#ggaed599bf26e2baa98c2d108a072d02337a21147eba9db31edf831bc27da595c909',1,'uart1_shared.h']]],
+  ['rxd_5fp30_5ftxd_5fp31_11',['RxD_P30_TxD_P31',['../group__uart1__shared.html#ggaed599bf26e2baa98c2d108a072d02337abb894b0d9862cbfcd47c453dba0b9c47',1,'uart1_shared.h']]],
+  ['rxd_5fp36_5ftxd_5fp37_12',['RxD_P36_TxD_P37',['../group__uart1__shared.html#ggaed599bf26e2baa98c2d108a072d02337a5d9bd9050fee81f933458cb44d02eed3',1,'uart1_shared.h']]]
 ];