+++ docs/html/search/all_1a.js
@@ -17,8 +17,10 @@ var searchData=
   ['what_20it_20is_14',['What it is',['../index.html#autotoc_md1',1,'']]],
   ['wirc_5fh_5faddress_15',['WIRC_H_ADDRESS',['../group__power__management.html#gacae7914e25fb794eb0d55b0162f8f8f6',1,'power_management.h']]],
   ['wirc_5fl_5faddress_16',['WIRC_L_ADDRESS',['../group__power__management.html#ga82e21440e9b642907b03b4f1d111ca7b',1,'power_management.h']]],
-  ['with_20parity_208_20bit_20with_20two_20stop_20bits_20send_20receive_20functions_17',['UART1 9bit, 8-bit with parity, 8-bit with two stop bits send/receive functions',['../group__uart1__9bit__shared.html',1,'']]],
-  ['with_20timer2_2012t_18',['with timer2 12t',['../group__uart1__mode1__timer2__12T.html',1,'UART1 Mode 1 with Timer2 12T'],['../group__uart1__mode3__timer2__12T.html',1,'UART1 Mode 3 with Timer2 12T']]],
-  ['with_20timer2_201t_19',['with timer2 1t',['../group__uart1__mode1__timer2__1T.html',1,'UART1 Mode 1 with Timer2 1T'],['../group__uart1__mode3__timer2__1T.html',1,'UART1 Mode 3 with Timer2 1T']]],
-  ['write_5fop_20',['WRITE_OP',['../group__eeprom.html#gga02a1ca0f05ee0d6aa0f1b6c2fc51af94a673cb8cebf7f88741d345dc00cba0c3d',1,'eeprom.h']]]
+  ['with_20cmake_17',['with cmake',['../md_utils_2memory__dumper_2README.html#autotoc_md18',1,'With CMake'],['../md_utils_2memory__dumper_2README.html#autotoc_md21',1,'With CMake']]],
+  ['with_20parity_208_20bit_20with_20two_20stop_20bits_20send_20receive_20functions_18',['UART1 9bit, 8-bit with parity, 8-bit with two stop bits send/receive functions',['../group__uart1__9bit__shared.html',1,'']]],
+  ['with_20platformio_20scons_19',['With PlatformIO / SCons',['../md_utils_2memory__dumper_2README.html#autotoc_md19',1,'']]],
+  ['with_20timer2_2012t_20',['with timer2 12t',['../group__uart1__mode1__timer2__12T.html',1,'UART1 Mode 1 with Timer2 12T'],['../group__uart1__mode3__timer2__12T.html',1,'UART1 Mode 3 with Timer2 12T']]],
+  ['with_20timer2_201t_21',['with timer2 1t',['../group__uart1__mode1__timer2__1T.html',1,'UART1 Mode 1 with Timer2 1T'],['../group__uart1__mode3__timer2__1T.html',1,'UART1 Mode 3 with Timer2 1T']]],
+  ['write_5fop_22',['WRITE_OP',['../group__eeprom.html#gga02a1ca0f05ee0d6aa0f1b6c2fc51af94a673cb8cebf7f88741d345dc00cba0c3d',1,'eeprom.h']]]
 ];