+++ docs/html/search/all_b.js
@@ -1,11 +1,14 @@
 var searchData=
 [
-  ['freq_5fhz_0',['freq_hz',['../timer2__mode0__Hz__to__ticks__example_8c.html#ac07f8b71ab75b4b60f9e7303aa98689b',1,'FREQ_HZ:&#160;timer2_mode0_Hz_to_ticks_example.c'],['../timer0__mode1__Hz__to__ticks__example_8c.html#ac07f8b71ab75b4b60f9e7303aa98689b',1,'FREQ_HZ:&#160;timer0_mode1_Hz_to_ticks_example.c'],['../timer0__mode2__Hz__to__ticks__example_8c.html#ac07f8b71ab75b4b60f9e7303aa98689b',1,'FREQ_HZ:&#160;timer0_mode2_Hz_to_ticks_example.c'],['../timer0__mode3__Hz__to__ticks__example_8c.html#ac07f8b71ab75b4b60f9e7303aa98689b',1,'FREQ_HZ:&#160;timer0_mode3_Hz_to_ticks_example.c']]],
-  ['freq_5fkhz_1',['FREQ_KHZ',['../timer0__mode0__Hz__to__ticks__example_8c.html#a7ae67445b01c3460da57a317e732b186',1,'timer0_mode0_Hz_to_ticks_example.c']]],
-  ['frequency_2',['Frequency',['../group__freq.html',1,'']]],
-  ['frequency_20and_20vise_20versa_20conversions_3',['Timer ticks to duration/frequency and vise versa conversions',['../group__timer__to__ms.html',1,'']]],
-  ['frequency_20convertions_4',['frequency convertions',['../group__timer0__mode0__to__ms.html',1,'Timer0 Mode0 ticks to duration/frequency convertions'],['../group__timer0__mode1__to__ms.html',1,'Timer0 Mode1 ticks to duration/frequency convertions'],['../group__timer0__mode2__to__ms.html',1,'Timer0 Mode2 ticks to duration/frequency convertions'],['../group__timer0__mode3__to__ms.html',1,'Timer0 Mode3 ticks to duration/frequency convertions'],['../group__timer2__mode0__to__ms.html',1,'Timer2 Mode0 ticks to duration/frequency convertions']]],
-  ['frequency_2eh_5',['frequency.h',['../frequency_8h.html',1,'']]],
-  ['from_20terminal_20on_20linux_6',['How to run unit tests from terminal on Linux',['../md_utils_2csv2hex_2README.html#autotoc_md13',1,'']]],
-  ['functions_7',['functions',['../group__uart1__8bit__shared.html',1,'UART1 8-bit send/receive functions'],['../group__uart1__9bit__shared.html',1,'UART1 9bit, 8-bit with parity, 8-bit with two stop bits send/receive functions'],['../index.html#autotoc_md5',1,'Use STC15 HAL functions']]]
+  ['files_0',['Files',['../md_utils_2memory__dumper_2README.html#autotoc_md22',1,'']]],
+  ['flashing_1',['Flashing',['../md_utils_2memory__dumper_2README.html#autotoc_md20',1,'']]],
+  ['format_2',['Output Format',['../md_utils_2memory__dumper_2README.html#autotoc_md16',1,'']]],
+  ['freq_5fhz_3',['freq_hz',['../timer2__mode0__Hz__to__ticks__example_8c.html#ac07f8b71ab75b4b60f9e7303aa98689b',1,'FREQ_HZ:&#160;timer2_mode0_Hz_to_ticks_example.c'],['../timer0__mode3__Hz__to__ticks__example_8c.html#ac07f8b71ab75b4b60f9e7303aa98689b',1,'FREQ_HZ:&#160;timer0_mode3_Hz_to_ticks_example.c'],['../timer0__mode2__Hz__to__ticks__example_8c.html#ac07f8b71ab75b4b60f9e7303aa98689b',1,'FREQ_HZ:&#160;timer0_mode2_Hz_to_ticks_example.c'],['../timer0__mode1__Hz__to__ticks__example_8c.html#ac07f8b71ab75b4b60f9e7303aa98689b',1,'FREQ_HZ:&#160;timer0_mode1_Hz_to_ticks_example.c']]],
+  ['freq_5fkhz_4',['FREQ_KHZ',['../timer0__mode0__Hz__to__ticks__example_8c.html#a7ae67445b01c3460da57a317e732b186',1,'timer0_mode0_Hz_to_ticks_example.c']]],
+  ['frequency_5',['Frequency',['../group__freq.html',1,'']]],
+  ['frequency_20and_20vise_20versa_20conversions_6',['Timer ticks to duration/frequency and vise versa conversions',['../group__timer__to__ms.html',1,'']]],
+  ['frequency_20convertions_7',['frequency convertions',['../group__timer0__mode0__to__ms.html',1,'Timer0 Mode0 ticks to duration/frequency convertions'],['../group__timer0__mode1__to__ms.html',1,'Timer0 Mode1 ticks to duration/frequency convertions'],['../group__timer0__mode2__to__ms.html',1,'Timer0 Mode2 ticks to duration/frequency convertions'],['../group__timer0__mode3__to__ms.html',1,'Timer0 Mode3 ticks to duration/frequency convertions'],['../group__timer2__mode0__to__ms.html',1,'Timer2 Mode0 ticks to duration/frequency convertions']]],
+  ['frequency_2eh_8',['frequency.h',['../frequency_8h.html',1,'']]],
+  ['from_20terminal_20on_20linux_9',['How to run unit tests from terminal on Linux',['../md_utils_2csv2hex_2README.html#autotoc_md13',1,'']]],
+  ['functions_10',['functions',['../group__uart1__8bit__shared.html',1,'UART1 8-bit send/receive functions'],['../group__uart1__9bit__shared.html',1,'UART1 9bit, 8-bit with parity, 8-bit with two stop bits send/receive functions'],['../index.html#autotoc_md5',1,'Use STC15 HAL functions']]]
 ];