+++ docs/html/search/files_7.js
@@ -1,4 +1,6 @@
 var searchData=
 [
-  ['power_5fmanagement_2eh_0',['power_management.h',['../power__management_8h.html',1,'']]]
+  ['mem_5fdump_2ec_0',['mem_dump.c',['../mem__dump_8c.html',1,'']]],
+  ['mem_5fprint_2ec_1',['mem_print.c',['../mem__print_8c.html',1,'']]],
+  ['mem_5fprint_2eh_2',['mem_print.h',['../mem__print_8h.html',1,'']]]
 ];