+++ docs/html/search/functions_4.js
@@ -1,4 +1,9 @@
 var searchData=
 [
-  ['putchar_0',['putchar',['../group__uart.html#ga07036e93939040c563e862a6a31b65bc',1,'putchar(int ch):&#160;uart.c'],['../group__uart.html#ga07036e93939040c563e862a6a31b65bc',1,'putchar(int ch):&#160;uart.c']]]
+  ['print_5fbyte_5fvalue_0',['print_byte_value',['../mem__print_8c.html#a95f323cd0902309bd206209f531d1f16',1,'print_byte_value(uint8_t byte):&#160;mem_print.c'],['../mem__print_8h.html#a95f323cd0902309bd206209f531d1f16',1,'print_byte_value(uint8_t byte):&#160;mem_print.c']]],
+  ['print_5ffooter_5frow_1',['print_footer_row',['../mem__print_8c.html#ad211d2741222dfc76c3324c3c82dce2e',1,'print_footer_row(void):&#160;mem_print.c'],['../mem__print_8h.html#ad211d2741222dfc76c3324c3c82dce2e',1,'print_footer_row(void):&#160;mem_print.c']]],
+  ['print_5fheader_5frow_2',['print_header_row',['../mem__print_8c.html#a002026cf7978cce9adf5d5a3d38bee48',1,'print_header_row(void):&#160;mem_print.c'],['../mem__print_8h.html#a002026cf7978cce9adf5d5a3d38bee48',1,'print_header_row(void):&#160;mem_print.c']]],
+  ['print_5frow_5fend_3',['print_row_end',['../mem__print_8c.html#a985addff1528e73f4a9ba989666f3cf1',1,'print_row_end(void):&#160;mem_print.c'],['../mem__print_8h.html#a985addff1528e73f4a9ba989666f3cf1',1,'print_row_end(void):&#160;mem_print.c']]],
+  ['print_5frow_5foffset_4',['print_row_offset',['../mem__print_8c.html#ab0e1bfbb95c9e3204b7b0252716988de',1,'print_row_offset(uint8_t address):&#160;mem_print.c'],['../mem__print_8h.html#ab0e1bfbb95c9e3204b7b0252716988de',1,'print_row_offset(uint8_t address):&#160;mem_print.c']]],
+  ['putchar_5',['putchar',['../group__uart.html#ga07036e93939040c563e862a6a31b65bc',1,'putchar(int ch):&#160;uart.c'],['../group__uart.html#ga07036e93939040c563e862a6a31b65bc',1,'putchar(int ch):&#160;uart.c']]]
 ];