+++ docs/html/search/pages_1.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['dumper_0',['Memory dumper',['../md_utils_2memory__dumper_2README.html',1,'']]]
+  ['dumper_0',['Memory Dumper',['../md_utils_2memory__dumper_2README.html',1,'']]]
 ];