+++ docs/html/search/pages_3.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['memory_20dumper_0',['Memory dumper',['../md_utils_2memory__dumper_2README.html',1,'']]]
+  ['memory_20dumper_0',['Memory Dumper',['../md_utils_2memory__dumper_2README.html',1,'']]]
 ];