+++ CHANGELOG.md
@@ -195,7 +195,7 @@
 * Timer0 sync delay routines implemented
 * CMake examples flash_* targets are fixed
 * Memory dump not only idata but xdata memory too
-* TODO Add docs to memory dumper README
+* Memory dumper is documented 
 * TODO Add docs to csv2hex README
 * TODO Make firmware sizes table on build
 * TODO Rework bits module